pyproject.toml
[project]
name = "gamble"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "hypothesis>=6.152.4",
    "matplotlib>=3.10.9",
    "numpy>=2.4.4",
    "pandas>=3.0.2",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.2",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
]
